Remove git rev list from CommandRunner's reponsibility
