+++ src/main/java/reposense/system/CommandRunner.java
@@ -3,9 +3,6 @@ package reposense.system;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Date;
-
-import reposense.git.Util;
 
 public class CommandRunner {
 
@@ -17,21 +14,6 @@ public class CommandRunner {
                 rootPath,
                 "java -jar checkstyle-7.7-all.jar -c /google_checks.xml -f xml " + absoluteDirectory
         );
-    }
-
-    /**
-     * Returns the latest commit hash before {@code date}.
-     * Returns an empty {@code String} if {@code date} is null, or there is no such commit.
-     */
-    public static String getCommitHashBeforeDate(String root, String branchName, Date date) {
-        if (date == null) {
-            return "";
-        }
-
-        Path rootPath = Paths.get(root);
-        String revListCommand = "git rev-list -1 --before="
-                + Util.GIT_LOG_SINCE_DATE_FORMAT.format(date) + " " + branchName;
-        return runCommand(rootPath, revListCommand);
     }
 
     public static String runCommand(Path path, String command) {
