+++ src/test/java/reposense/git/GitRevListTest.java
@@ -1,4 +1,4 @@
-package reposense.system;
+package reposense.git;
 
 import java.util.Calendar;
 import java.util.Date;
@@ -9,19 +9,19 @@ import org.junit.Test;
 import reposense.template.GitTestTemplate;
 import reposense.util.TestUtil;
 
-public class CommandRunnerTest extends GitTestTemplate {
+public class GitRevListTest extends GitTestTemplate {
 
     @Test
     public void getCommitHashBeforeDate_beforeInitialCommitDate_emptyResult() {
         Date date = TestUtil.getDate(2018, Calendar.FEBRUARY, 4);
-        String commitHash = CommandRunner.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), date);
+        String commitHash = GitRevList.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), date);
         Assert.assertTrue(commitHash.isEmpty());
     }
 
     @Test
     public void getCommitHashBeforeDate_afterLatestCommitDate_success() {
         Date date = TestUtil.getDate(2018, Calendar.MAY, 10);
-        String commitHash = CommandRunner.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), date);
+        String commitHash = GitRevList.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), date);
 
         // result from git has a newline at the end
         Assert.assertEquals(EUGENE_AUTHOR_README_FILE_COMMIT_07052018 + "\n", commitHash);
@@ -30,7 +30,7 @@ public class CommandRunnerTest extends GitTestTemplate {
     @Test
     public void getCommitHashBeforeDate_februaryNineDate_success() {
         Date date = TestUtil.getDate(2018, Calendar.FEBRUARY, 9);
-        String commitHash = CommandRunner.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), date);
+        String commitHash = GitRevList.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), date);
 
         // result from git has a newline at the end
         Assert.assertEquals(FAKE_AUTHOR_BLAME_TEST_FILE_COMMIT_08022018 + "\n", commitHash);
@@ -38,13 +38,13 @@ public class CommandRunnerTest extends GitTestTemplate {
 
     @Test
     public void getCommitHashBeforeDate_nullDate_emptyResult() {
-        String commitHash = CommandRunner.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), null);
+        String commitHash = GitRevList.getCommitHashBeforeDate(config.getRepoRoot(), config.getBranch(), null);
         Assert.assertTrue(commitHash.isEmpty());
     }
 
     @Test(expected = RuntimeException.class)
     public void getCommitHashBeforeDate_invalidBranch_throwsRunTimeException() {
         Date date = TestUtil.getDate(2018, Calendar.FEBRUARY, 9);
-        CommandRunner.getCommitHashBeforeDate(config.getRepoRoot(), "invalidBranch", date);
+        GitRevList.getCommitHashBeforeDate(config.getRepoRoot(), "invalidBranch", date);
     }
 }
